+++ ECourseWeb/src/main/java/com/htt/pojo/Tag.java
@@ -40,7 +40,7 @@ public class Tag implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 100)
@@ -53,20 +53,20 @@ public class Tag implements Serializable {
     public Tag() {
     }
 
-    public Tag(Integer id) {
+    public Tag(Long id) {
         this.id = id;
     }
 
-    public Tag(Integer id, String name) {
+    public Tag(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
